Move apriltags and vision to their own crates
