+++ libs/vv-apriltag/src/data/34139872896_defdb2f8d9_c.jpg
