+++ libs/vv-utils/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "vv-utils"
+version.workspace = true
+license.workspace = true
+edition.workspace = true
+
+[dependencies]
+serde = { workspace = true, optional = true }
+
+[features]
+serde = ["dep:serde"]
