+++ libs/vv-vision/src/buffer/conv/tests.rs
