+++ libs/vv-vision/src/vision/data/ferris.png
