+++ src/lib.rs
@@ -1,16 +1,11 @@
-pub mod apriltag;
-pub mod broadcast;
-pub mod buffer;
 pub mod camera;
 pub mod component_filter;
 pub mod components;
-pub mod draw;
-pub mod geom;
 pub mod pipeline;
 pub mod serialized;
 pub mod utils;
-pub mod vision;
-pub mod vision_debug;
+
+pub use vv_vision::*;
 
 /// Mutexes to be used throughout the crate
 ///
